kill first
If yarn map reduce heavy job is submitted from shib, such a job can't be submitted soon as map reduce job.
It takes a long time to submit a job as map reduce.

As the result, If you giveup the query, although shib DB is only update, job is not killed.

So, kill map reduce job and then update shib DB.